+++ packages/expo-clipboard/src/web/ClipboardModule.ts
@@ -18,6 +18,7 @@ import {
   GetImageOptions,
   GetStringOptions,
   SetStringOptions,
+  SetImageOptions,
   StringFormat,
 } from '../Clipboard.types';
 
@@ -153,7 +154,7 @@ export default {
       throw new PasteFailureException(error.message);
     }
   },
-  async setImageAsync(base64image: string): Promise<void> {
+  async setImageAsync(base64image: string, _options?: SetImageOptions): Promise<void> {
     if (!navigator.clipboard) {
       throw new ClipboardUnavailableException();
     }